Modify deprecated imresize code in observation
Searching the solutions about that, I find that using cv2.resize works in this case.
So I edit it.
The code works !!

Note that the resize takes (width, height) as the resize parameter instead of (height, width, channe;) <-- the shape in code,
Followed that by setting the second argument to (self.shape[1], self.shape[0]) <-- (width, height).