Change filename of saving models and way of flattening layer
We change the filename format by adding the total timestep in the filename.

Also, as x.view(x.size(0), -1) can't be run at colab,
we change the way of coding it to x.reshape(),
since it is same as the former and able to run in colab.

Update README.md